made necessary changes
