+++ Turtle Graphics/README.md
@@ -1,0 +1,13 @@
+This repository consists of a list of python scripts to automate few tasks.
+
+You can contribute by adding more python scripts which can be used to automate things. Some of already done are listed below.
+Incase you have anything to be followed while executing the python script mention it as well
+
+
+# Python Script
+
+## Script  - Turtle Graphics
+
+Code using turtle graphics
+TurtleGraphics.py
+
